Frontend Finale

+++ Frontend/src/app/components/book-list/book-list.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { Book } from 'src/app/models/Book';
 import { OperationResponse } from 'src/app/models/OperationResponse';
 import { BookService } from 'src/app/services/book.service';
+
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-book-list',
@@ -10,54 +13,65 @@ import { BookService } from 'src/app/services/book.service';
 })
 export class BookListComponent implements OnInit {
 
+  actualPage: number = 1;
   bookList: Book[] = [];
-  loading: boolean = true;
+  loading: boolean = false;
+  bookFilter: any = { id: '' };
 
-  constructor(private _bookService: BookService) { }
+  constructor(private _bookService: BookService, private router: Router) { }
 
   ngOnInit(): void {
     this.getAllBooks();
   }
 
   getAllBooks() {
+    this.loading = true;
     this._bookService.getBooks().subscribe(d => {
-      console.log(d)
+      //console.log(d)
+      if (d.success == false) {
+        Swal.fire(
+          'There has been an error!',
+          d.message,
+          'error'
+        )
+      }
+      this.bookList = d.data!;
       this.loading = false;
     }, err => {
-      console.log(err.message);
-    })
-  }
-
-  getBook(id: number) {
-    this._bookService.getBook(id).subscribe(b => {
-      console.log(b);
-    }, err => {
-      console.log(err.message);
-    })
-  }
-
-  addBook(book: Book) {
-    this._bookService.addBook(book).subscribe(b => {
-      console.log(b);
-    }, err => {
-      console.log(err.message);
+      Swal.fire(
+        'There has been an error!',
+        err.message,
+        'error'
+      )
     })
   }
 
   deleteBook(id: number) {
-    this._bookService.deleteBook(id).subscribe(b => {
-      console.log(b);
-    }, err => {
-      console.log(err.message);
+    Swal.fire({
+      title: 'Are you sure you want to delete?',
+      showCancelButton: true,
+      confirmButtonText: 'Yes',
+      denyButtonText: `No`,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._bookService.deleteBook(id).subscribe(b => {
+          Swal.fire(
+            'Deleted!',
+            b.message,
+            'success'
+          )
+          this.getAllBooks();
+        }, err => {
+          Swal.fire(
+            'There has been an error!',
+            err.message,
+            'error'
+          )
+        })
+      }
     })
-  }
 
-  updateBook(id: number, book: Book) {
-    this._bookService.updateBook(id, book).subscribe(b => {
-      console.log(`id: ${id}, book: ${book}`);
-    }, err => {
-      console.log(err.message);
-    })
+
   }
